Update mdance src for release 0.2.5
Included data files under `src/mdance/data` directory with MD data and
2D data, expected test outputs. Some docstrings were updated for
clarity.
