+++ src/mdance/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.2.2'
+__version__ = '0.2.5'
 __author__ = 'Lexin Chen'
 __author_email__ = 'le.chen@ufl.edu'
 __description__ = 'a flexible n-ary clustering package for molecular dynamics trajectories.'
