+++ src/mdance/data/__init__.py
@@ -1,0 +1,15 @@
+from __future__ import absolute_import
+
+import importlib.resources
+
+
+top = importlib.resources.files(__name__) / 'aligned_tau.pdb'
+traj = importlib.resources.files(__name__) / 'aligned_1000_tau.dcd'
+sim_traj_numpy = importlib.resources.files(__name__) / 'backbone.npy'
+
+cc_sim = importlib.resources.files(__name__) / 'cc_sim.npy'
+trimmed_sim = importlib.resources.files(__name__) / 'trimmed_sim.npy'
+
+blob_disk = importlib.resources.files(__name__) / 'blob_disk.csv'
+diamond = importlib.resources.files(__name__) / 'diamond9.csv'
+ellipses = importlib.resources.files(__name__) / 'ellipses.csv'
