+++ src/mdance/data/ellipses.csv
@@ -1,0 +1,523 @@
+# Data_10_2 or AD_10_2: This data is in 2-d space, and has 10 clusters.  The total number of  points is 500.
+#
+# If you use this data, please quote the following references:
+#
+# S. Bandyopadhyay and U. Maulik, ``Nonparametric genetic clustering: Comparison validity indices'',
+# IEEE Transactions on Systems, Man and Cybernetics, Part C, vol. 31, no. 1, pp. 120-125, 2001
+#
+# S. Bandyopadhyay and U. Maulik, ``Genetic Clustering for Automatic Evolution
+# of Clusters and Application to Image Classification'',  Pattern Recognition, vol.35,
+# pp. 1197-1208, 2002
+#
+# S. Bandyopadhyay and S. K. Pal, ``Classification and Learning Using Genetic
+# Algorithms: Applications in Bioinformatics and Web Intelligence", Springer,
+# Heidelberg, 2007
+# 
+# source: http://www.isical.ac.in/~sanghami/data.html
+#RELATION elliptical_10_2
+
+#ATTRIBUTE x REAL
+#ATTRIBUTE y REAL
+#ATTRIBUTE class {0,1,2,3,4,5,6,7,8,9}
+
+#DATA
+3.51,8.23,4
+1.15,7.98,4
+0.36,9.5,4
+0.75,8.45,4
+0.73,8.27,4
+-0.1,7.94,4
+-0.18,6.3,4
+3.26,6.88,4
+-0.88,7.83,4
+-1.9,7.67,4
+2.69,8.71,4
+-0.27,8.91,4
+2.44,9.5,4
+-1.01,8.5,4
+-0.89,9.21,4
+3.21,7.17,4
+1.35,7.64,4
+-0.03,8.47,4
+3.28,7.86,4
+-0.01,7.68,4
+3.19,8.95,4
+-0.54,8.18,4
+0.39,7.13,4
+2.75,9.17,4
+0.08,7.1,4
+-0.42,7.61,4
+-1.24,7.36,4
+-1.87,8.33,4
+-1.38,8.78,4
+-0.83,8.1,4
+0.64,6.72,4
+-0.15,6.55,4
+0.36,8.75,4
+-0.63,6.76,4
+1.37,7.87,4
+-1.7,8.83,4
+2.58,8.09,4
+0.18,7.66,4
+-1.26,7.62,4
+1.37,9.06,4
+-0.09,6.72,4
+2.72,8.73,4
+-0.03,9.43,4
+2.38,6.71,4
+-0.5,9.21,4
+-1.18,6.92,4
+-1.68,7.51,4
+0.17,9.49,4
+1.28,7.82,4
+0.31,6.54,4
+6.23,3.1,5
+6.77,4.38,5
+7.09,3.27,5
+2.45,4.39,5
+6.59,4.56,5
+4.22,2.94,5
+3.66,5.42,5
+2.96,5,5
+5.64,4.64,5
+3.12,5.01,5
+5.39,5.66,5
+3.84,5.67,5
+3.47,2.65,5
+6.53,3.05,5
+4.3,3.71,5
+3.38,4.3,5
+5.03,2.61,5
+5.23,5.38,5
+7.11,3.54,5
+2.55,3.35,5
+6.39,2.91,5
+5.97,4.79,5
+3.57,3.75,5
+2.96,3.88,5
+5.74,5.3,5
+3.13,5.29,5
+6.22,3.43,5
+4.95,3.88,5
+5.42,2.29,5
+3.42,3.83,5
+4.3,4.15,5
+2.97,3.89,5
+2.73,4.98,5
+4.12,3.66,5
+6.13,3.26,5
+4.91,2.3,5
+2.9,5.06,5
+2.35,3.55,5
+7.36,4.95,5
+5.71,2.66,5
+6.56,4.28,5
+3.65,5.28,5
+6.44,2.52,5
+3.1,2.9,5
+3.72,3.89,5
+3.77,3.29,5
+6.1,2.2,5
+3.9,3.49,5
+4.12,4.97,5
+3.8,3.71,5
+0.74,-5.26,6
+-0.22,-6.19,6
+-1.44,-6.59,6
+-1.81,-6.49,6
+0.54,-5.04,6
+-3.06,-7.39,6
+0.14,-6.69,6
+-1.81,-7.64,6
+-3.01,-5.02,6
+-3.8,-4.65,6
+-0.96,-5.37,6
+-1.51,-4.37,6
+-1.28,-7.34,6
+-4.64,-5.19,6
+-2.34,-7.03,6
+-4.69,-6.23,6
+0.09,-6.91,6
+-1.14,-6.33,6
+-1.91,-4.32,6
+-0.82,-4.39,6
+-4.51,-5.64,6
+-3.06,-4.23,6
+0.8,-6.32,6
+-1.87,-5.2,6
+-1.5,-6.57,6
+-0.79,-4.58,6
+-0.38,-5.96,6
+-4.15,-6.54,6
+-3.4,-7.2,6
+-0.32,-6.29,6
+-2.73,-4.26,6
+-3.34,-5.03,6
+0.1,-4.95,6
+-1.17,-4.98,6
+-2.93,-5.28,6
+-3.71,-6.66,6
+-0.27,-7.33,6
+-3.05,-7.13,6
+-3.03,-7.27,6
+-1.8,-7.12,6
+-2.04,-4.33,6
+-2.72,-4.49,6
+-3.25,-5.96,6
+-3.86,-7.34,6
+-0.34,-6.89,6
+0.28,-7.09,6
+-3.83,-7.1,6
+-2.24,-5.96,6
+-0.44,-7.1,6
+-3.06,-4.57,6
+0.57,1.59,2
+-1.1,0.14,2
+-0.46,-1.79,2
+2.72,-0.34,2
+0.97,-0.3,2
+0.45,-1.17,2
+-1.24,-0.78,2
+1.82,0.34,2
+0.86,1.69,2
+-0.76,-1.92,2
+0.26,-0.42,2
+1.93,-1.25,2
+0.45,0.1,2
+-2,0.1,2
+2.24,0.82,2
+-1.12,1.84,2
+1.67,0.82,2
+-0.64,-1.36,2
+-1.17,1.43,2
+0.24,-1.9,2
+1.26,-1.43,2
+-0.68,0.03,2
+-2.25,-1.11,2
+-0.6,-1.72,2
+1.25,0.68,2
+-0.71,-1.38,2
+0.61,-1.83,2
+-1.27,0.98,2
+-1.78,-0.14,2
+0.26,-1.05,2
+1.49,1.3,2
+-2.67,-0.53,2
+1.22,-1.52,2
+-0.76,1.62,2
+-0.26,-0.74,2
+-0.64,0.39,2
+2.66,0.22,2
+-0.62,-0.75,2
+1.75,0.4,2
+2.69,-0.21,2
+0.6,-1.84,2
+-0.19,1.94,2
+-2.72,0.76,2
+-2.34,0.73,2
+-0.08,-0.06,2
+-0.71,0.75,2
+1.62,0.3,2
+0.03,1.58,2
+-1.34,0.95,2
+-1.93,1.32,2
+8.4,0.95,1
+12.67,-0.43,1
+11.25,-1.81,1
+11.74,-0.54,1
+11.83,0.55,1
+10.2,-0.81,1
+8.66,-1.11,1
+9.32,-1.42,1
+10.37,-0.59,1
+9.23,0.43,1
+11.26,-1.32,1
+12.18,-0.46,1
+9.15,-1.48,1
+9.94,0.34,1
+11.77,-0.88,1
+11.81,0.31,1
+13.06,0.02,1
+12.45,-0.7,1
+10.91,-0.86,1
+8.85,-0.51,1
+10.34,0.15,1
+9.99,-0.31,1
+11.09,1.02,1
+13.07,1.08,1
+10.23,-1.75,1
+9.04,1.05,1
+12.25,-0.38,1
+9.91,0.62,1
+9.13,0.24,1
+11.69,0.28,1
+12.58,0.69,1
+9.23,-0.66,1
+10.49,-0.05,1
+12.34,-0.87,1
+8.89,0.78,1
+8.07,0.3,1
+12,0.8,1
+12.82,-1.32,1
+10.06,-0.04,1
+10.62,0.68,1
+9.72,0.32,1
+10.6,-1.34,1
+8.38,0.14,1
+12.29,0.89,1
+10.05,-1.03,1
+11.3,1.54,1
+10.07,0.98,1
+10.07,-1.09,1
+13.66,-0.87,1
+8.46,1.07,1
+-11.64,1.37,9
+-10.71,-0.95,9
+-12.22,0.62,9
+-10.64,0.57,9
+-10.93,0.41,9
+-11.43,0.62,9
+-13.04,1.65,9
+-9.74,-0.11,9
+-10.69,-1.03,9
+-10.95,-0.37,9
+-11.04,1.41,9
+-11.99,0.33,9
+-9.75,1.04,9
+-9.69,1.27,9
+-10.24,0.13,9
+-9.84,-0.18,9
+-11.56,-1.1,9
+-13.34,1.65,9
+-10.35,-1.05,9
+-11.92,1.42,9
+-13.35,-1.19,8
+-10.4,-0.58,9
+-11.87,-0.02,9
+-10.62,-1.66,9
+-10.73,-0.57,9
+-12.86,-0.97,8
+-10.38,-0.43,9
+-9.63,-1.21,9
+-12.67,1.68,9
+-10.83,0,9
+-11.1,0.29,9
+-9.78,-1.14,9
+-9.8,1.11,9
+-10.97,0.73,9
+-9.59,-0.87,9
+-12.25,1.35,9
+-11.72,-1.82,9
+-9.69,0.46,9
+-10.09,-1.03,9
+-12.67,-1.27,8
+-9.43,-0.86,9
+-12.62,0.78,9
+-10.74,0.53,9
+-10.74,-1.52,9
+-9.07,0.08,9
+-13.53,-0.73,8
+-10.37,-1.61,9
+-12.09,1.71,9
+-11.46,0.73,9
+-9.31,0.03,9
+7.18,9.04,3
+5.82,8.79,3
+6.97,11.86,3
+5.75,9.03,3
+6.76,9.48,3
+9.28,10.41,3
+8.22,11.69,3
+6.59,8.8,3
+8.38,8.71,3
+5.71,9.23,3
+9.12,10.29,3
+10.22,10.54,3
+6.78,10.07,3
+6.6,10.11,3
+8.04,8.51,3
+6.1,9.34,3
+10.67,9.58,3
+6.43,9.5,3
+5.38,9.55,3
+6.54,10.07,3
+5.82,10.07,3
+10.18,9.18,3
+5.36,10.12,3
+7.81,8.68,3
+5.62,9.1,3
+8.48,8.95,3
+7.79,10.12,3
+9.36,9.48,3
+8.27,9.74,3
+10,10.04,3
+5.5,10.49,3
+9.76,8.47,3
+7.58,11.9,3
+10.5,10.88,3
+10.23,9.9,3
+7.6,11.41,3
+6.87,9.26,3
+8.05,8.02,3
+6.38,11.31,3
+6.59,8.6,3
+5.89,9.83,3
+10.9,9.58,3
+10.84,10.42,3
+10.08,9.68,3
+8.17,10.86,3
+10.35,9.36,3
+6.01,9.17,3
+9.6,11.25,3
+6.51,9.79,3
+9.25,8.77,3
+12.4,3.31,0
+15.47,2.23,0
+15.06,1.62,0
+12.71,3.93,0
+17.59,2.61,0
+13.96,3.05,0
+17.51,2.37,0
+13.47,3.63,0
+16.56,3.13,0
+13.99,1.71,0
+17.04,1.81,0
+14.35,4.39,0
+16.72,3.98,0
+16.15,1.92,0
+17.87,2.99,0
+13.09,2.1,0
+17.26,2.16,0
+15.78,4.6,0
+13,2.13,0
+13.77,1.46,0
+14.89,2.92,0
+17.43,3.72,0
+17.77,2.69,0
+16.98,2.6,0
+17.39,3.18,0
+14.65,4.43,0
+12.7,3.55,0
+15.68,3.58,0
+13.03,2.59,0
+16.16,4.09,0
+13.04,2.97,0
+15.04,1.1,0
+13.68,3.41,0
+16.19,4.36,0
+16.9,3.12,0
+16.3,3.44,0
+14.04,4.03,0
+16.58,2.88,0
+16.22,4.05,0
+13.37,1.83,0
+13.34,1.34,0
+17,2.2,0
+16.35,3.61,0
+16.19,1.98,0
+13.54,4.16,0
+16.45,1.48,0
+16.57,2.9,0
+17.15,1.95,0
+16.5,3.1,0
+13.45,4.22,0
+-17.15,-1.94,8
+-15.84,-4.15,8
+-15.42,-1.77,8
+-15.88,-4.2,8
+-16.26,-3.35,8
+-13.71,-3.24,8
+-15.19,-1.71,8
+-17.64,-3.9,8
+-15.82,-3.47,8
+-13.85,-4.61,8
+-13.01,-3.64,8
+-15.98,-3.49,8
+-17.27,-4.29,8
+-15.09,-3.65,8
+-16.18,-2.44,8
+-16.13,-1.47,8
+-12.34,-2.87,8
+-14.55,-4.11,8
+-16.81,-3.41,8
+-15.8,-3.63,8
+-14.44,-1.84,8
+-17.78,-3.31,8
+-14.9,-4.27,8
+-17.2,-4.32,8
+-13.01,-3.99,8
+-13.74,-2.83,8
+-12.18,-2.92,8
+-14.23,-2.26,8
+-16.51,-3.37,8
+-13.42,-2.78,8
+-14.31,-2.18,8
+-13.63,-4.24,8
+-16.18,-3.5,8
+-14.46,-2.32,8
+-15.08,-4,8
+-16.1,-3.55,8
+-15.69,-1.1,8
+-17.39,-3.1,8
+-15.85,-1.94,8
+-15.74,-3.54,8
+-14.06,-1.3,8
+-15.92,-4.38,8
+-12.78,-2.33,8
+-13.12,-3.82,8
+-12.94,-3.46,8
+-17.66,-2.42,8
+-13.37,-1.88,8
+-14.38,-2.81,8
+-16.39,-2.04,8
+-13.94,-4.29,8
+-5.1,2.07,7
+-9.66,2.24,7
+-6.91,2.66,7
+-7.19,3.22,7
+-5.2,2.76,7
+-7.76,1.88,7
+-4.65,4.2,7
+-4.36,2.86,7
+-7.5,4.16,7
+-9.62,2.82,7
+-7.37,2.25,7
+-4.58,3.41,7
+-7.8,1.62,7
+-9.26,2.93,7
+-9.11,2.78,7
+-5.27,3.19,7
+-4.36,3.25,7
+-4.55,2.85,7
+-9.05,2.72,7
+-7.46,1.85,7
+-7.71,2.2,7
+-4.48,2.9,7
+-8.91,3.16,7
+-8.7,4.34,7
+-9.77,3.49,7
+-8.36,3.63,7
+-9.3,2.03,7
+-6.03,3.55,7
+-5.5,2.07,7
+-8.12,2.91,7
+-6.83,2.15,7
+-5.19,2.13,7
+-8.37,1.31,7
+-7.15,4.51,7
+-7.81,3.17,7
+-7.97,3.8,7
+-7.31,1.38,7
+-5.07,3.48,7
+-5.3,3.19,7
+-9.83,2.83,7
+-7.1,5,7
+-4.93,3,7
+-5.28,4.25,7
+-6.12,1.61,7
+-6.61,1.17,7
+-6.19,2.09,7
+-8.57,3.76,7
+-8.44,2.62,7
+-7.33,1.8,7
+-6.85,4.63,7
