+++ src/mdance/prime/sim_calc.py
@@ -3,7 +3,7 @@ import json
 import numpy as np
 import re
 
-from mdance.tools.esim_modules import SimilarityIndex, calc_medoid, calc_outlier
+from mdance.tools.esim import SimilarityIndex, calc_medoid, calc_outlier
 
 
 class FrameSimilarity:
